feat(wear): Implement watch-side data caching and pruning
This commit introduces a robust local caching and data management strategy on the watch to ensure no steps are lost during periods of disconnection.

- **Step Caching**: A `StepCacheWorker` has been implemented to periodically save the unsynced step count to the local Room database every 15 minutes.
- **Data Pruning**: A `StepPruningWorker` has been added to delete cached step records older than 48 hours, preventing the local database from growing indefinitely.
- **Worker Scheduling**: Both workers are scheduled on application startup using `WorkManager` in the `WearApplication` class.
- **Refactoring**:
    - The `StepTrackingService` was updated to expose the `unsyncedStepCount` as a `StateFlow` for safe access from background workers.
    - The `CachedStepEntity` `stepDelta` property was changed from `Int` to `Long` to ensure type consistency.

+++ stepblockswear/src/main/java/com/stepblocks/wear/WearableDataListenerService.kt
@@ -39,7 +39,7 @@ class WearableDataListenerService : WearableListenerService() {
                     val dataMapArrayList = ArrayList<DataMap>()
                     historicalData.forEach {
                         val data = DataMap()
-                        data.putInt("step_delta", it.stepDelta)
+                        data.putLong("step_delta", it.stepDelta)
                         data.putLong("timestamp", it.timestamp)
                         dataMapArrayList.add(data)
                     }
